Update seed.ts
Move randomX functions to new file in ~/utils.

Adjust number of items to be generated.
